Added a way to use the display without needing to acquire any locks (for panic); Also redesigned the page allocator a bit (mostly the initialization routine, to be able to use early KeFatalError()s on most of it)

+++ src/boot/osloader/include/memory.h
@@ -31,8 +31,8 @@
 typedef struct {
     RtDList ListHeader;
     uint8_t Type;
-    uint32_t BasePage;
-    uint32_t PageCount;
+    uint64_t BasePage;
+    uint64_t PageCount;
 } OslpMemoryDescriptor;
 
 typedef struct OslpMemoryArenaEntry {
